Add files via upload
was literally missing a space in one of my audio file names and that caused the whole menu not to function.  Huge shout out to professor Adam for reminding me about the developer tool as I was able to find that out thanks to the error message.  But now the actual play scene isn't working... However, now that I have the tool, and see the error, I should be able to fix it